% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{zibppca}
\alias{zibppca}
\title{Principal component analysis for Zero-inflated Bivariate Poisson variables}
\usage{
zibppca(data, scale. = TRUE, maxit = 30, pres = 1e-06)
}
\arguments{
\item{data}{a data frame, providing the data for the principal components analysis}

\item{scale.}{a logical value, indicating whether the variables should be scaled to have unit variance before the analysis takes place. The default is FALSE.}

\item{maxit}{a scalar, number of maximum iterations}

\item{pres}{a scalar, the precision of the EM algorithm}
}
\description{
This function preforms principal component analysis for pair-wise 
Zero-inflated Bivariate Poisson distributed variables.
}
